+++ c++/sym-ISOLR.cc
@@ -19,7 +19,7 @@ class SymmetryISOLRcommon : public SymLR {
   }
 
   // Multiplicity of the I=(II,SS,P) subspace = (2I+1)(2S+1) = II SS.
-  int mult(const Invar &I) override {
+  size_t mult(const Invar &I) override {
     int mi = I.get("II"); // isospin multiplicity
     int ms = I.get("SS"); // spin multiplicity
     return mi * ms;
