+++ c++/sym-QS.cc
@@ -19,7 +19,7 @@ class SymmetryQS : public Symmetry {
   }
 
   // Multiplicity of the (Q,SS) subspace is 2S+1 = SS.
-  int mult(const Invar &I) override { return I.get("SS"); }
+  size_t mult(const Invar &I) override { return I.get("SS"); }
 
   bool triangle_inequality(const Invar &I1, const Invar &I2, const Invar &I3) override {
     return u1_equality(I1.get("Q"), I2.get("Q"), I3.get("Q")) && su2_triangle_inequality(I1.get("SS"), I2.get("SS"), I3.get("SS"));
